Make 'qemu-vexpress-a9' and 'imx6ul' use the same libcpu code.
