+++ bsp/imx6ul/platform/SConscript
@@ -8,7 +8,6 @@ drivers/imx_timer.c
 drivers/imx_i2c.c
 drivers/imx_uart.c
 cpu/armv7_cache.c
-cpu/gic.c
 cpu/ccm_pll2.c
 cpu/mmu.c
 cpu/cortex_a_gcc.S
