+++ libcpu/arm/cortex-a/gic.h
@@ -15,6 +15,9 @@
 #ifndef __GIC_H__
 #define __GIC_H__
 
+#include <rthw.h>
+#include <platform.h>
+
 int arm_gic_dist_init(rt_uint32_t index, rt_uint32_t dist_base, int irq_start);
 int arm_gic_cpu_init(rt_uint32_t index, rt_uint32_t cpu_base);
 
@@ -26,8 +29,11 @@ void arm_gic_set_group(rt_uint32_t index, int vector, int group);
 int arm_gic_get_active_irq(rt_uint32_t index);
 void arm_gic_ack(rt_uint32_t index, int irq);
 
+void arm_gic_clear_active(rt_uint32_t index, int irq);
+void arm_gic_clear_pending(rt_uint32_t index, int irq);
+
 void arm_gic_dump_type(rt_uint32_t index);
-void rt_hw_vector_init(void);
+void arm_gic_dump(rt_uint32_t index);
 
 #endif
