+++ libcpu/arm/cortex-a/start_gcc.S
@@ -6,7 +6,11 @@
  * Change Logs:
  * Date           Author       Notes
  * 2013-07-05     Bernard      the first version
+ * 2018-11-22     Jesven       in the interrupt context, use rt_scheduler_do_irq_switch checks
+ *                             and switches to a new thread
  */
+
+#include "rtconfig.h"
 
 .equ Mode_USR,        0x10
 .equ Mode_FIQ,        0x11
@@ -20,11 +24,11 @@
 .equ F_Bit,           0x40            @ when F bit is set, FIQ is disabled
 
 .equ UND_Stack_Size,     0x00000000
-.equ SVC_Stack_Size,     0x00000100
+.equ SVC_Stack_Size,     0x00000400
 .equ ABT_Stack_Size,     0x00000000
 .equ RT_FIQ_STACK_PGSZ,  0x00000000
-.equ RT_IRQ_STACK_PGSZ,  0x00000100
-.equ USR_Stack_Size,     0x00000100
+.equ RT_IRQ_STACK_PGSZ,  0x00000800
+.equ USR_Stack_Size,     0x00000400
 
 #define ISR_Stack_Size  (UND_Stack_Size + SVC_Stack_Size + ABT_Stack_Size + \
                  RT_FIQ_STACK_PGSZ + RT_IRQ_STACK_PGSZ)
@@ -44,12 +48,24 @@ stack_top:
 /* reset entry */
 .globl _reset
 _reset:
-    bl      rt_cpu_mmu_disable
     /* set the cpu to SVC32 mode and disable interrupt */
-    mrs     r0, cpsr
-    bic     r0, r0, #0x1f
-    orr     r0, r0, #0x13
-    msr     cpsr_c, r0
+    cps #Mode_SVC
+    ldr sp, =svc_stack_2_limit
+    bl flush_cache_all
+    bl rt_cpu_mmu_disable
+
+#ifdef RT_USING_SMP
+    mrc p15, 0, r1, c1, c0, 1
+    mov r0, #(1<<6)
+    orr r1, r0
+    mcr p15, 0, r1, c1, c0, 1 //enable smp
+#endif
+
+    ldr lr, =after_enable_mmu
+    ldr r0, =mtbl
+    b enable_mmu
+
+after_enable_mmu:
 
     /* setup stack */
     bl      stack_setup
@@ -114,6 +130,85 @@ stack_setup:
     msr     cpsr_c, #Mode_SVC|I_Bit|F_Bit
     bx      lr
 
+    .global enable_mmu
+enable_mmu:
+    orr r0, #0x18
+    mcr p15, 0, r0, c2, c0, 0 @ttbr0
+
+    mov r0, #(1 << 5)         @PD1=1
+    mcr p15, 0, r0, c2, c0, 2 @ttbcr
+
+    mov r0, #1
+    mcr p15, 0, r0, c3, c0, 0 @dacr
+
+    mov r0, #0
+    mcr p15, 0, r0, c8, c7, 0
+    mcr p15, 0, r0, c7, c5, 0 @iciallu
+    mcr p15, 0, r0, c7, c5, 6 @bpiall
+
+    mrc p15, 0, r0, c1, c0, 0
+    orr r0, #(1 | 4)
+    orr r0, #(1 << 12)
+    mcr p15, 0, r0, c1, c0, 0
+    dsb
+    isb
+    mov pc, lr
+
+.global flush_cache_all
+flush_cache_all:
+    stmfd   sp!, {r0-r12, lr}
+    bl  v7_flush_dcache_all
+    mov r0, #0
+    mcr p15, 0, r0, c7, c5, 0       @ I+BTB cache invalidate
+    dsb
+    isb
+    ldmfd   sp!, {r0-r12, lr}
+    mov pc, lr
+
+    v7_flush_dcache_all:
+    dmb                 @ ensure ordering with previous memory accesses
+    mrc p15, 1, r0, c0, c0, 1       @ read clidr
+    ands    r3, r0, #0x7000000      @ extract loc from clidr
+    mov r3, r3, lsr #23         @ left align loc bit field
+    beq finished            @ if loc is 0, then no need to clean
+    mov r10, #0             @ start clean at cache level 0
+loop1:
+    add r2, r10, r10, lsr #1        @ work out 3x current cache level
+    mov r1, r0, lsr r2          @ extract cache type bits from clidr
+    and r1, r1, #7          @ mask of the bits for current cache only
+    cmp r1, #2              @ see what cache we have at this level
+    blt skip                @ skip if no cache, or just i-cache
+    mcr p15, 2, r10, c0, c0, 0      @ select current cache level in cssr
+    isb                 @ isb to sych the new cssr&csidr
+    mrc p15, 1, r1, c0, c0, 0       @ read the new csidr
+    and r2, r1, #7          @ extract the length of the cache lines
+    add r2, r2, #4          @ add 4 (line length offset)
+    ldr r4, =0x3ff
+    ands    r4, r4, r1, lsr #3      @ find maximum number on the way size
+    clz r5, r4              @ find bit position of way size increment
+    ldr r7, =0x7fff
+    ands    r7, r7, r1, lsr #13     @ extract max number of the index size
+loop2:
+    mov r9, r4              @ create working copy of max way size
+loop3:
+    orr r11, r10, r9, lsl r5        @ factor way and cache number into r11
+    orr r11, r11, r7, lsl r2        @ factor index number into r11
+    mcr p15, 0, r11, c7, c14, 2     @ clean & invalidate by set/way
+    subs    r9, r9, #1          @ decrement the way
+    bge loop3
+    subs    r7, r7, #1          @ decrement the index
+    bge loop2
+skip:
+    add r10, r10, #2            @ increment cache number
+    cmp r3, r10
+    bgt loop1
+finished:
+    mov r10, #0             @ swith back to cache level 0
+    mcr p15, 2, r10, c0, c0, 0      @ select current cache level in cssr
+    dsb
+    isb
+    mov pc, lr
+
 /* exception handlers: undef, swi, padt, dabt, resv, irq, fiq          */
 .section .text.isr, "ax"
     .align  5
@@ -137,12 +232,22 @@ vector_fiq:
     .align  5
 .globl vector_irq
 vector_irq:
+#ifdef RT_USING_SMP
+    clrex
+#endif
     stmfd   sp!, {r0-r12,lr}
 
     bl      rt_interrupt_enter
     bl      rt_hw_trap_irq
     bl      rt_interrupt_leave
 
+#ifdef RT_USING_SMP
+    mov     r0, sp
+    bl      rt_scheduler_do_irq_switch
+
+    ldmfd   sp!, {r0-r12,lr}
+    subs    pc,  lr, #4
+#else
     @ if rt_thread_switch_interrupt_flag set, jump to
     @ rt_hw_context_switch_interrupt_do and don't return
     ldr     r0, =rt_thread_switch_interrupt_flag
@@ -174,6 +279,11 @@ rt_hw_context_switch_interrupt_do:
     stmfd   sp!, {r1-r4}    @ push old task's r0-r3
     stmfd   sp!, {r0}       @ push old task's cpsr
 
+#ifdef RT_USING_LWP
+    stmfd sp, {r13, r14}^  @push usr_sp, usr_lr
+    sub sp, #8
+#endif
+
     ldr     r4,  =rt_interrupt_from_thread
     ldr     r5,  [r4]
     str     sp,  [r5]       @ store sp in preempted tasks's TCB
@@ -182,10 +292,17 @@ rt_hw_context_switch_interrupt_do:
     ldr     r6,  [r6]
     ldr     sp,  [r6]       @ get new task's stack pointer
 
+#ifdef RT_USING_LWP
+    ldmfd sp, {r13, r14}^  @pop usr_sp, usr_lr
+    add sp, #8
+#endif
+
     ldmfd   sp!, {r4}       @ pop new task's cpsr to spsr
     msr     spsr_cxsf, r4
 
     ldmfd   sp!, {r0-r12,lr,pc}^ @ pop new task's r0-r12,lr & pc, copy spsr to cpsr
+
+#endif
 
 .macro push_svc_reg
     sub     sp, sp, #17 * 4         @/* Sizeof(struct rt_hw_exp_stack)  */
@@ -200,36 +317,90 @@ rt_hw_context_switch_interrupt_do:
 .endm
 
     .align  5
-    .globl	vector_swi
+    .globl  vector_swi
+.weak SVC_Handler
+SVC_Handler:
 vector_swi:
     push_svc_reg
     bl      rt_hw_trap_swi
     b       .
 
     .align  5
-    .globl	vector_undef
+    .globl  vector_undef
 vector_undef:
     push_svc_reg
     bl      rt_hw_trap_undef
     b       .
 
     .align  5
-    .globl	vector_pabt
+    .globl  vector_pabt
 vector_pabt:
     push_svc_reg
     bl      rt_hw_trap_pabt
     b       .
 
     .align  5
-    .globl	vector_dabt
+    .globl  vector_dabt
 vector_dabt:
     push_svc_reg
     bl      rt_hw_trap_dabt
     b       .
 
     .align  5
-    .globl	vector_resv
+    .globl  vector_resv
 vector_resv:
     push_svc_reg
     bl      rt_hw_trap_resv
     b       .
+
+#ifdef RT_USING_SMP
+.global set_secondary_cpu_boot_address
+set_secondary_cpu_boot_address:
+    ldr r0, =secondary_cpu_start
+
+    mvn r1, #0 //0xffffffff
+    ldr r2, =0x10000034
+    str r1, [r2]
+    str r0, [r2, #-4]
+    mov pc, lr
+
+.global secondary_cpu_start
+secondary_cpu_start:
+    mrc p15, 0, r1, c1, c0, 1
+    mov r0, #(1<<6)
+    orr r1, r0
+    mcr p15, 0, r1, c1, c0, 1 //enable smp
+
+    ldr r0, =mtbl
+    ldr lr, =1f
+
+    b enable_mmu
+1:
+    mrc p15, 0, r0, c1, c0, 0
+    bic r0, #(1<<13)
+    mcr p15, 0, r0, c1, c0, 0
+
+    cps #Mode_IRQ
+    ldr sp, =irq_stack_2_limit
+
+    cps #Mode_FIQ
+    ldr sp, =irq_stack_2_limit
+
+    cps #Mode_SVC
+    ldr sp, =svc_stack_2_limit
+
+    b secondary_cpu_c_start
+#endif
+
+.bss
+.align 2   //align to  2~2=4
+svc_stack_2:
+    .space (1 << 10)
+svc_stack_2_limit:
+
+irq_stack_2:
+    .space (1 << 10)
+irq_stack_2_limit:
+
+/* include the mmu table defined for the specific platform */
+#include "mtbl.S"
